+++ perseo-quality/src/perseo_quality/spectral_analysis/support.py
@@ -5,14 +5,19 @@
 
 from __future__ import annotations
 
+from pathlib import Path
+
 import numpy as np
 from arepytools.timing.precisedatetime import PreciseDateTime
+from netCDF4 import Dataset
 from numpy.polynomial import Polynomial
 from scipy.constants import speed_of_light
 from scipy.fft import fft2, fftshift, ifft2, ifftshift
-from scipy.signal import spectrogram
+from scipy.ndimage import gaussian_filter1d
+from scipy.signal import find_peaks, spectrogram
 from scipy.signal.windows import hamming
 
+from perseo_quality import __version__
 from perseo_quality.core.generic_dataclasses import SARAcquisitionMode
 from perseo_quality.core.signal_processing import (
     convert_to_db,
@@ -20,6 +25,10 @@ from perseo_quality.core.signal_processing import (
     locate_max_2d_interp,
 )
 from perseo_quality.io.quality_input_protocol import ChannelData, SARCoordinatesFunction
+from perseo_quality.spectral_analysis.custom_dataclasses import (
+    DistributedSpectraDataOutput,
+    PointTargetSpectraDataOutput,
+)
 
 # TODO: check what can be moved into Perseo Core
 
@@ -53,7 +62,7 @@ def data_deramping(
     mid_burst_az_time, _ = channel_data.get_mid_burst_times(burst=burst)
     lines_step = channel_data.azimuth_axis[1] - channel_data.azimuth_axis[0]
 
-    # default: scansar mode
+    # default are for scansar mode
     azimuth_steering_rate_axis = None
     sensor_velocity_norm_mid_burst = None
     if channel_data.acquisition_mode == SARAcquisitionMode.TOPSAR:
@@ -75,8 +84,8 @@ def data_deramping(
         doppler_centroid_poly=channel_data.doppler_centroid,
         doppler_rate_poly=channel_data.doppler_rate,
         azimuth_steering_rate_axis=azimuth_steering_rate_axis,
-        carrier_frequency=channel_data.carrier_frequency,
         sensor_velocity_norm_mid_burst=sensor_velocity_norm_mid_burst,
+        wavelength=speed_of_light / channel_data.carrier_frequency,
     )
     burst_az_relative_target_location_px = roi_center_location_px[0] - burst_start_time_px
     burst_target_start_az_roi_crop = burst_az_relative_target_location_px - np.floor(data.shape[1] / 2).astype(int)
@@ -89,58 +98,22 @@ def data_deramping(
 
 
 def compute_deramping_phase_exponential(
-    lines_per_burst: int, lines_step: float, doppler_centroid_axis: np.ndarray, doppler_rate_axis: np.ndarray
-) -> np.ndarray:
-    """Computing deramping exponential phase function, ideally for scansar mode (without introducing the steering rate).
-
-    .. math::
-
-        exp \\left( \\pi i \\cdot k_D \\left(t_{burst}^{rel} - \\left(t_{mid\\_burst}^{rel} - \\frac{f_{DC}}{k_D}
-        \\right) \\right)^2 \\right)
-
-    where :math:`k_D` is the doppler rate axis and :math:`f_{DC}` is the doppler centroid axis.
-
-    Parameters
-    ----------
-    lines_per_burst : int
-        lines per burst
-    lines_step : float
-        lines step in seconds
-    doppler_centroid_axis : np.ndarray
-        doppler centroid frequency axis, computed for each range sample, with shape (samples,)
-    doppler_rate_axis : np.ndarray
-        doppler rate axis, computed for each range sample, with shape (samples,)
-
-    Returns
-    -------
-    np.ndarray
-        deramping exponential phase function, with shape (lines per burst, samples)
-    """
-    burst_deramping_azimuth_time_axis = compute_deramping_azimuth_axis(
-        lines_per_burst=lines_per_burst,
-        lines_step=lines_step,
-        doppler_centroid_axis=doppler_centroid_axis,
-        doppler_rate_axis=doppler_rate_axis,
-    )
-    deramping_phase = np.pi * doppler_rate_axis * (burst_deramping_azimuth_time_axis) ** 2
-    return np.exp(1j * deramping_phase)
-
-
-def compute_deramping_phase_exponential_with_steering(
     lines_per_burst: int,
     lines_step: float,
-    sensor_velocity_norm_mid_burst: float,
-    carrier_frequency: float,
     doppler_centroid_axis: np.ndarray,
     doppler_rate_axis: np.ndarray,
-    azimuth_steering_rate_axis: np.ndarray,
+    steering_rate_factor: np.ndarray | None = None,
 ) -> np.ndarray:
-    """Computing deramping exponential phase function, ideally for topsar mode, using the azimuth steering rate.
+    """Computing deramping exponential phase function. It can be used both for Scansar and Topsar acquisition modes,
+    providing the azimuth steering rate for Topsar mode.
 
     .. math::
 
-        exp \\left( \\pi i \\cdot \\frac{-k_D \\cdot a_r}{a_r - k_D} \\left(t_{burst}^{rel} -
-        \\left(t_{mid\\_burst}^{rel} - \\frac{f_{DC}}{k_D} \\right) \\right)^2 \\right)
+        \\begin{align}exp \\left( \\pi i \\cdot coeff \\cdot \\left(t_{burst}^{rel} - \\left(t_{mid\\_burst}^{rel} -
+        \\frac{(f_{DC} - f_{DC_{mid}})}{k_D} \\right) \\right)^2 \\right)\\
+        coeff_{topsar} = -\\frac{-k_D \\cdot a_r}{a_r - k_D}\\
+        coeff_{scansar} = k_D\\
+        \\end{align}
 
     Parameters
     ----------
@@ -164,21 +137,17 @@ def compute_deramping_phase_exponential_with_steering(
     np.ndarray
         deramping exponential phase function, with shape (lines per burst, samples)
     """
-
     burst_deramping_azimuth_time_axis = compute_deramping_azimuth_axis(
         lines_per_burst=lines_per_burst,
         lines_step=lines_step,
         doppler_centroid_axis=doppler_centroid_axis,
         doppler_rate_axis=doppler_rate_axis,
     )
-    wavelength = speed_of_light / carrier_frequency
-    steering_rate_factor = 2 * sensor_velocity_norm_mid_burst / wavelength * azimuth_steering_rate_axis
-    deramping_phase = (
-        -np.pi
-        * (-doppler_rate_axis * steering_rate_factor)
-        / (steering_rate_factor - doppler_rate_axis)
-        * (burst_deramping_azimuth_time_axis) ** 2
-    )
+    coeff = doppler_rate_axis
+    if steering_rate_factor is not None:
+        # TOPSAR deramping steering rate K_TU factor
+        coeff = -(-doppler_rate_axis * steering_rate_factor) / (steering_rate_factor - doppler_rate_axis)
+    deramping_phase = np.pi * coeff * (burst_deramping_azimuth_time_axis) ** 2
 
     return np.exp(1j * deramping_phase)
 
@@ -187,6 +156,10 @@ def compute_deramping_azimuth_axis(
     lines_per_burst: int, lines_step: float, doppler_centroid_axis: np.ndarray, doppler_rate_axis: np.ndarray
 ) -> np.ndarray:
     """Computing deramping phase burst azimuth relative axis.
+
+    .. math::
+
+        t_{burst}^{rel} - \\left(t_{mid\\_burst}^{rel} - \\frac{(f_{DC} - f_{DC_{mid}})}{k_D}\\right)
 
     Parameters
     ----------
@@ -206,7 +179,9 @@ def compute_deramping_azimuth_axis(
     """
 
     burst_azimuth_relative_axis = np.arange(0, lines_per_burst, 1) * lines_step
-    beam_center_times = -doppler_centroid_axis / doppler_rate_axis
+    beam_center_times = (
+        -(doppler_centroid_axis - doppler_centroid_axis[doppler_centroid_axis.size // 2]) / doppler_rate_axis
+    )
     reference_times = lines_per_burst / 2 * lines_step + beam_center_times
     return burst_azimuth_relative_axis.reshape(-1, 1) - reference_times
 
@@ -258,13 +233,13 @@ def compute_burst_deramping_function(
     doppler_rate_poly: SARCoordinatesFunction,
     azimuth_steering_rate_axis: np.ndarray | None = None,
     sensor_velocity_norm_mid_burst: float | None = None,
-    carrier_frequency: float | None = None,
+    wavelength: float | None = None,
 ) -> np.ndarray:
-    """Computing the deramping and demodulation phase exponential function for topsar and scansar acquisition for a
+    """Computing the deramping and demodulation phase exponential function for Topsar and Scansar acquisition for a
     given burst, defined by the mid burst azimuth time.
 
-    Azimuth steering rate is needed only for topsar deramping phase computation, it can be omitted for
-    scansar deramping.
+    Azimuth steering rate, Normalized sensor velocity and sensor wavelength are needed only for Topsar deramping phase
+    computation.
 
     Parameters
     ----------
@@ -281,7 +256,11 @@ def compute_burst_deramping_function(
     doppler_rate_poly : SARCoordinatesFunction
         doppler rate polynomial
     azimuth_steering_rate_axis : np.ndarray | None, optional
-        steering rate axis, needed for topsar deramping, by default None
+        steering rate axis, needed for Topsar deramping, by default None
+    sensor_velocity_norm_mid_burst : float | None, optional
+        normalized sensor velocity at mid burst, needed for Topsar deramping, by default None
+    wavelength : float | None, optional
+        sensor wavelength, needed for Topsar deramping, by default None
 
     Returns
     -------
@@ -295,27 +274,26 @@ def compute_burst_deramping_function(
     doppler_rate_axis = np.array(
         [doppler_rate_poly.evaluate(azimuth_time=mid_burst_az_time, range_time=r) for r in slant_range_axis]
     )
+    steering_rate_factor = None
     if azimuth_steering_rate_axis is not None:
         # topsar
-        if sensor_velocity_norm_mid_burst is None or carrier_frequency is None:
-            raise RuntimeError("For Topsar deramping steering rate, normalized velocity and fc_hz are needed")
-        deramping_phase_exp = compute_deramping_phase_exponential_with_steering(
-            lines_per_burst=lines_per_burst,
-            lines_step=lines_step,
+        if sensor_velocity_norm_mid_burst is None or wavelength is None:
+            raise RuntimeError(
+                "For Topsar deramping 'steering rate', 'normalized velocity' and 'wavelength' are needed"
+            )
+        steering_rate_factor = compute_steering_rate_factor(
+            wavelength=wavelength,
             sensor_velocity_norm_mid_burst=sensor_velocity_norm_mid_burst,
-            carrier_frequency=carrier_frequency,
-            doppler_centroid_axis=doppler_centroid_axis,
-            doppler_rate_axis=doppler_rate_axis,
             azimuth_steering_rate_axis=azimuth_steering_rate_axis,
         )
-    else:
-        # scansar
-        deramping_phase_exp = compute_deramping_phase_exponential(
-            lines_per_burst=lines_per_burst,
-            lines_step=lines_step,
-            doppler_centroid_axis=doppler_centroid_axis,
-            doppler_rate_axis=doppler_rate_axis,
-        )
+
+    deramping_phase_exp = compute_deramping_phase_exponential(
+        lines_per_burst=lines_per_burst,
+        lines_step=lines_step,
+        doppler_centroid_axis=doppler_centroid_axis,
+        doppler_rate_axis=doppler_rate_axis,
+        steering_rate_factor=steering_rate_factor,
+    )
 
     demodulation_phase_exp = compute_demodulation_phase_exponential(
         lines_per_burst=lines_per_burst,
@@ -325,6 +303,28 @@ def compute_burst_deramping_function(
     )
 
     return deramping_phase_exp * demodulation_phase_exp
+
+
+def compute_steering_rate_factor(
+    sensor_velocity_norm_mid_burst: float, wavelength: float, azimuth_steering_rate_axis: np.ndarray
+) -> np.ndarray:
+    """Computing steering rate factor for Topsar deramping.
+
+    Parameters
+    ----------
+    sensor_velocity_norm_mid_burst : float
+        normalized sensor velocity at mid burst
+    wavelength : float
+        sensor wavelength
+    azimuth_steering_rate_axis : np.ndarray
+        azimuth steering rate axis
+
+    Returns
+    -------
+    np.ndarray
+        steering rate factor for deramping
+    """
+    return 2 * sensor_velocity_norm_mid_burst / wavelength * azimuth_steering_rate_axis
 
 
 def recenter_data(data: np.ndarray) -> np.ndarray:
@@ -396,10 +396,10 @@ def compute_spectrogram_db(
     spectrogram_db = convert_to_db(
         np.abs((np.abs(near_spectrogram) + np.abs(mid_spectrogram) + np.abs(far_spectrogram)) / 3)
     )
-    return spectrogram_db, fftshift(spectrogram_freq), spectrogram_times
+    return fftshift(spectrogram_db, axes=0), fftshift(spectrogram_freq), spectrogram_times
 
 
-def compute_polynomial_fit(profile: np.ndarray, freq_axis: np.ndarray) -> Polynomial:
+def compute_polynomial_fit(profile: np.ndarray, freq_axis: np.ndarray, boundaries: tuple[float, float]) -> Polynomial:
     """FItting polynomial on profile portion.
 
     Parameters
@@ -408,15 +408,15 @@ def compute_polynomial_fit(profile: np.ndarray, freq_axis: np.ndarray) -> Polyno
         profile to be fit
     freq_axis : np.ndarray
         frequency axis
+    boundaries : tuple[float, float]
+        fitting region boundaries
 
     Returns
     -------
     Polynomial
         fitting Polynomial
     """
-    start_id = profile.size // 4
-    stop_id = -profile.size // 4 + 1
-    return Polynomial.fit(freq_axis[start_id:stop_id], profile[start_id:stop_id], deg=2)
+    return Polynomial.fit(freq_axis[boundaries[0] : boundaries[1]], profile[boundaries[0] : boundaries[1]], deg=2)
 
 
 def extract_abs_profiles(
@@ -477,6 +477,238 @@ def extract_phase_profiles(
         for start, stop in zip(az_profiles_starts, az_profiles_stops, strict=True)
     ]
     return range_profiles_deg, azimuth_profiles_deg
+
+
+def compute_spectrum_boundaries(profile: np.ndarray) -> tuple[int, int]:
+    """Computing relevant spectrum boundaries from profile.
+
+    Parameters
+    ----------
+    profile : np.ndarray
+        profile to be analyzed
+
+    Returns
+    -------
+    int
+        start index of spectrum
+    int
+        stop index of spectrum
+    """
+    profile_smooth = gaussian_filter1d(profile, sigma=3)
+    abs_derivative = np.abs(np.diff(profile_smooth))
+    peak_indexes, _ = find_peaks(abs_derivative / abs_derivative.max(), height=0.8, distance=50)
+    if peak_indexes.size < 2:
+        return 0, profile.size
+    start_idx = max([0, int(peak_indexes[0]) + 2])
+    stop_idx = min([profile.size, int(peak_indexes[-1]) - 2])
+    return start_idx, stop_idx
+
+
+def spectral_analysis_profiles_to_netcdf(
+    data: list[PointTargetSpectraDataOutput] | list[DistributedSpectraDataOutput],
+    out_path: str | Path,
+) -> Path:
+    """Saving Spectral Analysis Profiles output data to NetCDF4 file.
+
+    Hierarchy::
+
+    Point Target Spectral Analysis Hierarchy::
+
+        root/
+        ├── product_attributes...
+        └── swath
+            └── polarization
+                ├── channel_attributes...
+                ├── doppler_centroid [at target position]
+                ├── phase_value [at target position]
+                ├── azimuth_frequency_axis
+                ├── range_frequency_axis
+                ├── azimuth_profiles_abs
+                ├── range_profiles_abs
+                ├── azimuth_profiles_phase
+                ├── range_profiles_phase
+                ├── az_phase_polynomial_coefficients
+                └── rng_phase_polynomial_coefficients
+
+    Distributed Spectral Analysis Hierarchy::
+
+        root/
+        ├── product_attributes...
+        └── swath
+            └── polarization
+                ├── channel_attributes...
+                └── block
+                    ├── block_attributes...
+                    ├── azimuth_frequency_axis
+                    ├── range_frequency_axis
+                    ├── azimuth_profiles_abs
+                    └── range_profiles_abs
+
+    Parameters
+    ----------
+    data : list[PointTargetSpectraDataOutput] | list[DistributedSpectraDataOutput]
+        list of PointTargetSpectraDataOutput | DistributedSpectraDataOutput dataclasses, corresponding to the full
+        output of the radiometric analysis
+    out_path : str | Path
+        path where to save the NetCDF file
+
+
+    Returns
+    -------
+    Path
+        path to the output NetCDF file
+    """
+    is_pt_analysis = isinstance(data[0], PointTargetSpectraDataOutput)
+    tag = "pt" if is_pt_analysis else "distr"
+    out_name = tag + "_spectral_profiles_" + data[0].general_info.product
+    out_path = Path(out_path)
+    output_file = out_path.joinpath(out_name).with_suffix(".nc")
+
+    root = Dataset(output_file, "w", format="NETCDF4")
+    root.title = f"{'Point Target' if is_pt_analysis else 'Distributed'} Spectral Analysis"
+    root.history = f"Created by PERSEO Quality v{__version__}"
+    root.product = data[0].general_info.product
+    root.sensor = data[0].general_info.sensor
+    root.product_type = data[0].general_info.product_type
+    root.acquisition_mode = data[0].general_info.acquisition_mode
+    root.orbit_direction = data[0].general_info.orbit_direction
+
+    for item in data:
+        if item.general_info.swath not in root.groups:
+            swath_grp = root.createGroup(item.general_info.swath)
+        else:
+            swath_grp = root.groups[item.general_info.swath]
+        if item.general_info.polarization not in swath_grp.groups:
+            pol_grp = swath_grp.createGroup(item.general_info.polarization)
+        else:
+            pol_grp = swath_grp.groups[item.general_info.polarization]
+        pol_grp.swath = item.general_info.swath
+        pol_grp.channel = item.general_info.channel
+        pol_grp.polarization = item.general_info.polarization
+        pol_grp.acquisition_start_time = str(item.general_info.acquisition_start_time)
+
+        # creating common dimensions
+        if is_pt_analysis:
+            pol_grp.createDimension("targets", len(item.targets_info))
+            pol_grp.createDimension("azimuth", item.targets_info[0].azimuth_frequency_axis.size)
+            pol_grp.createDimension("range", item.targets_info[0].range_frequency_axis.size)
+            pol_grp.createDimension("slices", 3)
+            pol_grp.createDimension("coeffs", 3)
+
+            # name of analyzed targets
+            pol_grp.bursts = [n.burst for n in item.targets_info]
+            pol_grp.targets = [n.target_name for n in item.targets_info]
+
+            # doppler centroid and phase value at target position
+            dc_target = pol_grp.createVariable(
+                "doppler_centroid", item.targets_info[0].target_doppler_centroid_Hz.dtype, ("targets",)
+            )
+            dc_target.unit = "Hz"
+            dc_target.description = "Doppler centroid at the target position"
+            dc_target[:] = np.array([p.target_doppler_centroid_Hz for p in item.targets_info])
+            phase_target = pol_grp.createVariable(
+                "phase_value", item.targets_info[0].target_phase_value_deg.dtype, ("targets",)
+            )
+            phase_target.unit = "deg"
+            phase_target.description = "Phase value at the target position"
+            phase_target[:] = np.array([p.target_phase_value_deg for p in item.targets_info])
+
+            # frequency axes
+            az_freq_axis = pol_grp.createVariable(
+                "azimuth_frequency_axis", item.targets_info[0].azimuth_frequency_axis.dtype, ("azimuth",)
+            )
+            az_freq_axis.unit = "Hz"
+            az_freq_axis[:] = item.targets_info[0].azimuth_frequency_axis
+            rng_freq_axis = pol_grp.createVariable(
+                "range_frequency_axis", item.targets_info[0].range_frequency_axis.dtype, ("range",)
+            )
+            rng_freq_axis.unit = "Hz"
+            rng_freq_axis[:] = item.targets_info[0].range_frequency_axis
+
+            # absolute profiles
+            az_abs_profiles = pol_grp.createVariable(
+                "azimuth_profiles_abs",
+                item.targets_info[0].azimuth_profiles_db[0].dtype,
+                ("targets", "slices", "azimuth"),
+            )
+            az_abs_profiles.unit = "dB"
+            az_abs_profiles[:] = np.stack([p.azimuth_profiles_db for p in item.targets_info])
+            rng_abs_profiles = pol_grp.createVariable(
+                "range_profiles_abs", item.targets_info[0].range_profiles_db[0].dtype, ("targets", "slices", "range")
+            )
+            rng_abs_profiles.unit = "dB"
+            rng_abs_profiles[:] = np.stack([p.range_profiles_db for p in item.targets_info])
+
+            # phase profiles
+            az_phase_profiles = pol_grp.createVariable(
+                "azimuth_profiles_phase",
+                item.targets_info[0].azimuth_profiles_deg[0].dtype,
+                ("targets", "slices", "azimuth"),
+            )
+            az_phase_profiles.unit = "deg"
+            az_phase_profiles[:] = np.stack([p.azimuth_profiles_deg for p in item.targets_info])
+            rng_phase_profiles = pol_grp.createVariable(
+                "range_profiles_phase",
+                item.targets_info[0].range_profiles_deg[0].dtype,
+                ("targets", "slices", "range"),
+            )
+            rng_phase_profiles.unit = "deg"
+            rng_phase_profiles[:] = np.stack([p.range_profiles_deg for p in item.targets_info])
+
+            # polynomials
+            az_poly_coeffs = pol_grp.createVariable(
+                "az_phase_polynomial_coefficients",
+                item.targets_info[0].azimuth_polynomial_fit.convert().coef.dtype,
+                ("targets", "coeffs"),
+            )
+            az_poly_coeffs[:] = np.stack([p.azimuth_polynomial_fit.convert().coef for p in item.targets_info])
+            rng_poly_coeffs = pol_grp.createVariable(
+                "rng_phase_polynomial_coefficients",
+                item.targets_info[0].range_polynomial_fit.convert().coef.dtype,
+                ("targets", "coeffs"),
+            )
+            rng_poly_coeffs[:] = np.stack([p.range_polynomial_fit.convert().coef for p in item.targets_info])
+
+        else:
+            for block in item.blocks_info:
+                blk_grp = pol_grp.createGroup(f"block_{block.block_num}")
+                blk_grp.first_az_line_block = block.first_az_line_block
+                blk_grp.lines_block = block.lines_block
+                blk_grp.samples_block = block.samples_block
+                blk_grp.doppler_centroid_mid_block_Hz = block.doppler_centroid_mid_block
+
+                blk_grp.createDimension("azimuth", block.azimuth_frequency_axis.size)
+                blk_grp.createDimension("range", block.range_frequency_axis.size)
+                blk_grp.createDimension("slices", 3)
+
+                # frequency axes
+                az_freq_axis = blk_grp.createVariable(
+                    "azimuth_frequency_axis", block.azimuth_frequency_axis.dtype, ("azimuth",)
+                )
+                az_freq_axis.unit = "Hz"
+                az_freq_axis[:] = block.azimuth_frequency_axis
+                rng_freq_axis = blk_grp.createVariable(
+                    "range_frequency_axis", block.range_frequency_axis.dtype, ("range",)
+                )
+                rng_freq_axis.unit = "Hz"
+                rng_freq_axis[:] = block.range_frequency_axis
+
+                # absolute profiles
+                az_abs_profiles = blk_grp.createVariable(
+                    "azimuth_profiles_abs",
+                    block.azimuth_profiles_db[0].dtype,
+                    ("slices", "azimuth"),
+                )
+                az_abs_profiles.unit = "dB"
+                az_abs_profiles[:] = np.stack(block.azimuth_profiles_db)
+                rng_abs_profiles = blk_grp.createVariable(
+                    "range_profiles_abs", block.range_profiles_db[0].dtype, ("slices", "range")
+                )
+                rng_abs_profiles.unit = "dB"
+                rng_abs_profiles[:] = np.stack(block.range_profiles_db)
+
+        root.close()
+        return output_file
 
 
 def _frequency_axis_generation(freq_vect: np.ndarray, samples: int, prf: int = 1) -> np.ndarray: